+++ src/runtime/app.ts
@@ -6,7 +6,7 @@ import { createHooks, Hookable } from 'hookable'
 import { useRuntimeConfig } from './config'
 import { timingMiddleware } from './timing'
 import { cachedEventHandler } from './cache'
-import { createRouteRulesHandler, getRouteRulesForPath } from './route-options'
+import { createRouteRulesHandler, getRouteRulesForPath } from './route-rules'
 import { plugins } from '#internal/nitro/virtual/plugins'
 import errorHandler from '#internal/nitro/virtual/error-handler'
 import { handlers } from '#internal/nitro/virtual/server-handlers'
