+++ src/runtime/index.ts
@@ -4,4 +4,4 @@ export * from './cache'
 export { useNitroApp } from './app'
 export * from './plugin'
 export * from './renderer'
-export { getRouteRules } from './route-options'
+export { getRouteRules } from './route-rules'
